cluster: deprecate get_node_by_* functions
...in favor of the new get_node and get_nodes methods. Combined all
get_node_by_* checks into get_node and also added checks for other
unique identifiers. Updated get_node_by_* methods to prompt a
deprecation warning and then proxy the call to Cluster.get_node. The
get_node_by* methods will be removed in a future release.

The get_node function take an identifier string and returns a node if
it's able to match any unique instance attribute to the identifier. For
example, you can pass in the instance id, alias, dns name, public ip,
private ip, etc. The get_nodes method will do this for a list of
identifiers and also takes care to remove potential duplicates from the
resulting node list in case you, for example, pass an instance id and an
alias for the same node in the list of identifiers.