cluster: add num_nodes kwarg to remove_nodes
..and changed the 'nodes' arg to an optional kwarg. This makes it
possible to either pass in a list of nodes to remove as before OR pass
in a number of nodes to be automagically selected and removed. Nodes are
selected counting backwards from the highest node number.

Also changed the 'nodes' arg to Cluster.remove_node an optional kwarg -
it's now only needed if you want to remove a specific node otherwise the
node with the highest node number is fetched and removed.